+++ src/main/java/com/odintsov/wallpapers_project/initializers/dtos/PrintingJson.java
@@ -5,8 +5,8 @@ import java.util.List;
 public record PrintingJson(
         String name,
         String article,
-        Float basePrice,
-        Float salePrice,
+        Double basePrice,
+        Double salePrice,
         String image,
         String description,
         Integer quantity,
